[FIX] account_invoice_extract: remove useless field
The field extract_word_ids is in the account move view but raises an
access error when someone without invoice rights open the invoice
(rights you have when you have some sales rights).

As this field is not used in the view or in the associated JS, we can
remove it and fix the access rights issues.

closes odoo/enterprise#73173

X-original-commit: ecb8243845447d967f78431cefe4b726880280c6
Signed-off-by: Louis Baudoux (lba) <lba@odoo.com>
Signed-off-by: Pierre Masereel (pim) <pim@odoo.com>