[I18N] *l10n_{be,ch}_hr_*: fw-port translations
Manual fw-port of translations since they're not on transifex for saas
versions. To do this properly, the following was also done:
- re-export pot files
- fill in translations for new terms added in this version
- remove leftover translation files from localizations they didn't make
  sense for
- fixed typo in `l10n_be_social_balance_sheet`

Note that within saas-17.4: `l10n_ch_hr_payroll_elm` was merged into
`l10n_ch_hr_payroll`. Therefore the translations needed to be manually
ported over from the previous version's module into the new merged po
file. Because translations for IT and DE don't appear to be completed
for 17.0's localzations on transifex, these translations were not ported
over as it's a very painful manual job cross checking the po files of
the former version.

Additional modules for l10n_be_hr translations added as well in v18

closes odoo/enterprise#74371

X-original-commit: 430a7fac9e844d5f535cd036121c9b57d7a92e98
Signed-off-by: Louis Wicket (wil) <wil@odoo.com>
Signed-off-by: Tiffany Chang (tic) <tic@odoo.com>