[FIX] l10n_be_hr_payroll: notify 'Notified time off officers' only
The eDRS activity on time off request is sent to
the current logged in user if there are no
`Notified Time Off Officers` specified on the leave
type. However, this shouldn't be the case. The activity
shouldn't be created if there are no `Notified Time
Off Officers`.

This commit fixes this.

task-4267132

closes odoo/enterprise#73140

Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>