[IMP] stock_barcode: Generate EPC through API
Provides an EPC encoder and two endpoints to allow the generation of
EPC for RFID tags by the customer.

Currently the EPC Encoder only support SGTIN-96 and SGTIN-198 schemes.
However, it is designed with extensibility in mind to allow further
encoding in the future.

"/stock_barcode/get_epc" is a generic endpoint used to generate any kind
of EPC, it's based on the scheme name and a GS1 AI element string.
"/stock_barcode/get_epc_sgtin" is a dedicated endpoint to generate SGTIN
EPC for products, based on a unique GTIN and a list of tracking numbers.

There are two levels of exceptions during encoding:

- Exceptions on technical data occur when invalid data is provided for
technical fields (header, partition, filter). In this case, the base
structure (schema) to be used for encoding is unstable/corrupt, it's a
critical error for the current and further encoding process, we stop it
and prevent the user by throwing an exception.
- User data exceptions occur when invalid data is provided for value
holding fields. In this case, the schema is still safe, it's just that
the current field is corrupted, so we can't generate a valid EPC for the
current item, but further encoding may still be correct. We warn the
user by returning the error as the intended EPC string and continue
encoding for further items.

This fault tolerance allows the user to process a large amount of data
at once without being abruptly interrupted by a single small problem.

Task 4256189

Part-of: odoo/enterprise#73859
Signed-off-by: Quentin Wolfs (quwo) <quwo@odoo.com>