[FIX] l10n_br_edi: require less fields for services to individuals
It's not always required to send a full customer address for service
invoices [1] and users don't want to add unnecessary friction in their
sales process.

Although service invoice legislation happens on the level of the city,
all cities we know of operate this way. If they don't it's ok, Avatax
will then return the error instead of us.

opw-4281020

[1] For Sao Paulo: Artigo 84 e do parágrafo 3º do Decreto 53.151/2012:
https://legislacao.prefeitura.sp.gov.br/leis/decreto-53151-de-17-de-maio-de-2012

closes odoo/enterprise#73257

X-original-commit: 32e5650ba66719d38be7f18d8bd17d185aa8f222
Signed-off-by: Josse Colpaert (jco) <jco@odoo.com>
Signed-off-by: Joren Van Onder (jov) <jov@odoo.com>