[FIX] web_studio: do not consider virtual records when navigating X2manys
Have a ir.default that sets one line into a one2many on some models.
In studio, edit the form view of the main model, then navigate to edit the one2many, with the ir.default having been triggered (there is one virual line in the one2many).

Before this commit there was a crash because "false" is not a valid id for the staticlist.

After this commit, there is no crash, and the virtual record doesn't appear.

opw-4289233

closes odoo/enterprise#73278

X-original-commit: d242c7898305cc54590101df59c5d542e1671892
Signed-off-by: Lucas Perais (lpe) <lpe@odoo.com>