[FIX] web_studio: handle logic for *not* groups attribute
Have a node that goes:
```xml
<div groups="!module.some_group" />

```

Before this commit, this case was not handled and that not-group was just ignored.

This created confusion as the node, when the user was part of the group, would be invisible, the checkbox invisible
would be ticked, and uniticking it had no effect.

After this commit, we handle the case, with a second selector in the studio sidebar.

opw-4273853

closes odoo/enterprise#73458

X-original-commit: f72c6361979734f9ea8821d20ca4292fb99f2bc4
Signed-off-by: Bruno Boi (boi) <boi@odoo.com>
Signed-off-by: Lucas Perais (lpe) <lpe@odoo.com>