[FIX] account_iso20022: Don't report 'SEPA' as Service Level in ISO20022 files
'SEPA' used to be the default for SvcLvl files, which meant even non-SEPA files would have it set to 'SEPA', unless they explicitly overrode the function defining the value of the node. This commit fixes this, by not setting any value to SvcLvl by default, and overriding for SEPA payments.

X-original-commit: 4c04799f7747b6bc0400138ce28645c8ed3cbc64
Part-of: odoo/enterprise#73610
Signed-off-by: Wala Gauthier (gawa) <gawa@odoo.com>
Signed-off-by: Olivier Colson (oco) <oco@odoo.com>