[REF] account_iso20022: compute BIC tag depending on the payment method consistently
The CdtrAgt node did not compute the tag of its BIC node in the same way as the DbtrAgt node, calling the dedicated helper. Because of that, it required a bit more payment method-specific code in SEPA, so that BIC becomes BICFI in the latest version. This commit removes that and calls the appropriate helper everywhere.

X-original-commit: bb898df194f8cd524b0129b1c7fa7385c00985a5
Part-of: odoo/enterprise#73610
Signed-off-by: Wala Gauthier (gawa) <gawa@odoo.com>
Signed-off-by: Olivier Colson (oco) <oco@odoo.com>