[IMP] {sale_}planning,project_forecast: auto plan based on closest SOL/project
Currently, when auto plan schedules a shift linked to an SOL, it will try to keep all of the related shifts assigned to the same person.
However, if the customer renews their contract and a new SOL is created, auto plan won't assign the new shift to the same resource.
This doesn't make sense as it is likely for the customer to continue working with the person they are familiar with and who knows about their needs best.

After this commit:
- If only the SOL of a shift is set and auto plan is run, we plan it to the resource of the "closest" shift linked to an SOL of the same product and the same customer.
- If only the project of a shift is set and auto plan is run, we plan it to the resource of the "closest" shift linked to the same project.

In both cases, "closest" means we pick the shift having the closest deadline compared to the current date.
We also check if the resource is available. If not, we look at the next closest shift.

task-4295867

closes odoo/enterprise#73237

Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>