[IMP] account_reports: allow editing date filter values
Currently the new date filter widget on reports only allows users to
change a period by clicking the `<` and `>` arrows next to the period
description. This makes it cumbersome if a user want to switch from a
month in one year to the same month in another year. He will have to
click the arrow button at least 12 times.

In order to simplify this kind of operation, we allow users editing the
period text inline on clicking it after selection. That way they can
manually change the month, quarter, or year and the report will reload
with the correct period based on that.

Also, the date filters dropdown in the reports are totally not
accessible when using a keyboard.

This commit adds easy keyboard navigation to the date filter dropdown by
using:
* `left arrow` to go to the previous period
* `right arrow` to go to the next period
* `enter` to select an option and to enter or exit period edit mode
* `tab/up arrow` and `shift+tab/down arrow` to navigate between the
  items and inputs

task-4284292

closes odoo/enterprise#74992

Related: odoo/odoo#190141
Signed-off-by: Olivier Colson (oco) <oco@odoo.com>