[FIX] stock_barcode(_mrp): lot_id field
This commit fixes two issues with the `lot_id` field in the move line
form view in the Barcode app for the manufacturing operations.

1. The `lot_id` field was displayed when no product was selected.

2. The `lot_id` field was not displayed for by-product lines created at
   the creation of the MO.

The first issue was there because the compute fields to determine if we
want to hide `lot_id` and/or `lot_name` were false by default.
The second issue was there because those fields depend of the line's
picking code which depends of the line's picking type which depend of
the line's `production_id`. The issue is by-product lines has no
`production_id` set.

closes odoo/enterprise#73241

closes odoo/enterprise#73965

Signed-off-by: Quentin Wolfs (quwo) <quwo@odoo.com>
Signed-off-by: Steve Van Essche <svs@odoo.com>