[FIX] account: fix empty string and False/None inconsistencies
When empty, the name field of account.move.line could inconsistently hold either an empty string or False. This inconsistency led to different results when searching for lines with an empty label, as label->not set and label == "" yielded different outcomes. The latter behavior aligns with expected ORM operations.

This fix ensures the name field is set to False whenever the computed result is an empty string, standardizing search behavior. We also adapt some tests to the new behavior in enterprise.

opw-4167139

closes odoo/enterprise#82582

X-original-commit: f87e47a7382163e2d30c5beea2a249d54d980caa
Related: odoo/odoo#204193
Signed-off-by: William André (wan) <wan@odoo.com>