[IMP] documents: lazy load partners in share panel
In production, there is a delay of around 1 second between the moment
the user clicks on the share button and the share/permission panel appears.
That is not good UX.

To reduce that delay, this commit removes the fetching of partner options
(for members' invite) when the permission panel is about to open. The RPC
call is now done when users open the select menu to invite new members.

task-4309414

closes odoo/enterprise#73781

X-original-commit: 3d82f846bd3230b49af66a42f02bc9a1a2b18254
Signed-off-by: Florian Charlier (flch) <flch@odoo.com>