[FIX] hr_payroll: apply null safety check on timezone for fully flexible contract
**Issue:**
The client gets an error while accessing an employee's payslips if that employee has a contract based on attendances with a null allowed value for its working schedule.

**Expected:**
The client should be able to access the payslips regardless on the working schedule value, even blank.

**Steps to reproduce:**
- Activate Payroll app and presence based on attendances in employees' settings;
- Open or create a contract through an employee's file;
- Set "Work Entry Source" to "Attendances" and leave "Working Schedule" empty;
- Try to access the employee's payslips through the action button.

**Cause:**
No timezone found on a contract's calendar because the calendar is null.
https://github.com/odoo/enterprise/blob/18.0/hr_payroll/models/hr_payslip.py#L1141

**Fix:**
Add a timezone retrieving rule :
- The contract's calendar tz;
- The employee's tz;
- The company's tz;
- UTC.

**Linked:**
Community PR:
https://github.com/odoo/odoo/pull/186222

opw-4268672

closes odoo/enterprise#76040

X-original-commit: 4cba2fb58560a95d5ce93eb6e6c1483ef0428ba5
Related: odoo/odoo#191476
Signed-off-by: Pierre Lamotte (pila) <pila@odoo.com>