[IMP] timesheet_grid: remove redundant actions in timesheet form view
Before this commit, the `Validate` and `Reset to Draft` buttons are
displayed in the header of the timesheet form view but also in the
`CogMenu` (actions dropdown), which is a bit redundant since the buttons
in the header of the form view are sufficient.

This commit removes the `Validate` and `Reset to Draft` buttons inside
the Cog menu when we are in the form view.

task-4279748

Part-of: odoo/enterprise#73327
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>