[IMP] helpdesk: remove edit button in helpdesk ticket kanban view
Before this commit, the `Edit` button is displayed in the menu of the
kanban card of helpdesk ticket. This button is not really useful since
the user can easily go to the form view of ticket by clicking on the
kanban card.

This commit removes `Edit` button in the kanban card menu of helpdesk
ticket since that button is not really useful.

task-4262236

closes odoo/enterprise#73328

Signed-off-by: Audric Onockx (auon) <auon@odoo.com>