[FIX] documents: change the "My Documents" filter
Purpose
=======
During the migration, the owner can be reset, if the user does not have a write
access on the documents (because being owner gives all access on the record).

Because of that, the filter "My Documents" does not show most of our files on
next.odoo.com. To mitigate that, that filter now uses the contact in addition
to the owner, and during the migration, if we reset the owner, and if the
contact is False, we will set the contact to the partner of the owner.

Task-4310779

closes odoo/enterprise#73914

X-original-commit: b2d68ecebb3eefe62b961f7d34b0085ab51dcd7e
Signed-off-by: Patrick Hoste (pko) <pko@odoo.com>
Signed-off-by: Stéphane Debauche (std) <std@odoo.com>