[FIX] documents(_project): do not allow admin to see all documents by default
Purpose
=======
Create a new group, so admin don't see all document by default.

Copy in SUDO the folder when we copy a project, to avoid access issue (in case the
project manager has not enough rights to copy the folder). It wasn't catched before
because the test user was erp manager.

Adjust a test in documents_spreadsheet.

Task-4313355

closes odoo/enterprise#73708

X-original-commit: 861015dad0f4de7371e221cfac92b74e336f4fde
Signed-off-by: Patrick Hoste (pko) <pko@odoo.com>
Signed-off-by: Stéphane Debauche (std) <std@odoo.com>