[FIX] project_timesheet_forecast_sale: copy shifts linked to projects without allocated hours
Some clients were confused about not knowing why their shifts were not copied after setting a project to them (using the copy previous week feature).
The reason was that the project linked to their shifts had allocated hours equal to 0.
After this commit, if 0 hours are allocated to the project, we allow to copy the shift anyway to avoid confusion (as some clients are clearly not using this field).

task-4074736
version-17.0

closes odoo/enterprise#74439

X-original-commit: d3f85bcb38b4a55bdbe21eb86470b7c7bf1c8078
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>