[IMP] planning: remove redundant action in planning shift form view
Before this commit, `Reset to draft` action is displayed in the cog menu
of list and form view of planning.slot model. However, a button `Reset
To Draft` is also displayed in the header of the form view which is a
bit redundant to have the same action in 2 different spaces inside the
same view.

This commit removes `Reset To Draft` action displayed in the cog menu of
planning.slot form view to make sure the action is only displayed once
in the form view.

task-4279748

closes odoo/enterprise#73327

Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>