[FIX] account_sepa_direct_debit module description
The module description of account_sepa_direct_debit in the manifest
states that posting an invoice will automatically generate a payment.
Actually that feature was dropped in Odoo 13 with commit
e3d390c4455c620793c481874b4503fa91bb6125 but the `__manifest__.py` was
not updated at that time. This makes the situation uncomfortable where a
feature not present in the code is still advertised in the module
description.

This commit updates the description of the module in the manifest.

closes odoo/enterprise#74653

X-original-commit: c62fa8e4a4a8253f86629e62234a21a7082650b6
Related: odoo/odoo#188740
Signed-off-by: Wala Gauthier (gawa) <gawa@odoo.com>
Signed-off-by: Olivier Colson (oco) <oco@odoo.com>