[REF] *: remove comparison feature
The comparison menu is removed in https://github.com/odoo/odoo/pull/186454. Here
we simplify a bit the component MrpDisplay and the cohort view to no longer
accept a prop "comparison" (unused in the case of MrpDisplay).
Note that we do not thoroughly refactor the cohort view since there are ongoing
discussions on refactor it using the spreadsheet library.

Task ID: 4307873

closes odoo/enterprise#73467

Related: odoo/odoo#186454
Signed-off-by: Julien Carion (juca) <juca@odoo.com>