[FIX] sale_subscription(_stock): prevent _post_invoice_hook call in invoice _post
_post_invoice_hook was called on an empty RecordSet at the end of AccountMove._post when having the context "recurring_automatic".
This context is voluntarily put here to prevent _post_invoice_hook to be run after the invoice posting, so that it can be run at the end of the cron, when all subscriptions have been handled.

However, if _post_invoice_hook is called on an empty RecordSet, it will search for all subscriptions currently handled by the cron.

This often resulted in a situation where only the first subscription was ready for the hook, while the others are not, resulting in only the first delivery created.

OPW-4166852

closes odoo/enterprise#73573

X-original-commit: d35c801211453c504cd4d1e0136661da9406c56b
Signed-off-by: Arnaud Joset (arj) <arj@odoo.com>
Signed-off-by: David Fesquet (dafr) <dafr@odoo.com>