[FIX] web_studio: fix approval tour on apps module.
The trigger in the tour that filter apps was wrongly set to "Not Installed", instead of "Installed".

Indeed, that tour puts approval on ir.module's updgrade button, so on already installed apps, and check whether approval rules are triggered when clicking on the "upgrade" button. So we need to filter only installed apps in the ir.module kanban.

closes odoo/enterprise#73471

X-original-commit: 983cf77a80120ad05805f5dacce5da59e67663a6
Signed-off-by: Pierre Pulinckx (pipu) <pipu@odoo.com>