[FIX] l10n_cl_edi_pos: update invoice button checkbox
Issue:
In the Chilean POS localization, the checkbox contained in the invoice button is always checked.
This is confusing, since it seems the option is selected even when it isn't.

Cause:
By default, the checkbox depends on the order's "is_to_invoice" return value.
In the Chilean POS, the "to_invoice" value is set to true by default and never updated, so "is_to_invoice" always returns true.

Solution:
Change behaviour of the checkbox to use the same "highlightInvoiceButton" method as the button itself.

Ticket:
4298996

closes odoo/enterprise#74115

X-original-commit: 5c2743f1bf845f0d796c0228b217a2556152c362
Signed-off-by: Claire Bretton (clbr) <clbr@odoo.com>