[FIX] documents: restore scrollbar after preview
Purpose
=======
Fix the view scrollbar which isn't restored when the file preview is closed.

Specifications
==============
When a file is previewed, the view scrollbar is hidden so that
scrolling actions don't impact the view behind the file preview.
However when the preview is closed, the scrollbar isn't restored.
=> Calling the documentList onDeleteCallback method to correctly restore
the view and the scrollbar when closing the preview.

Task-4296803

closes odoo/enterprise#74358

X-original-commit: 2e58f9cdc0a89697fcaa4f75e1410a250eafcf08
Signed-off-by: Stéphane Debauche (std) <std@odoo.com>
Signed-off-by: Amélie Dieudonné (amdi) <amdi@odoo.com>