[FIX] hr_payroll: set payslip salary to 0 when no hours recorded
Steps to reproduce:
- Install Attendances
- New employee > New Contract > Set a wage
- Set 'Work Entry Source' to 'Attendances'
- Payroll app > New Payslip > Compute Sheet
- Salary Computation tab > Basic salary = contract Wage

Steps to check salary configurator:
- Install Salary Configurator and Recruitment
- Recruitment > Any Job Position > New Application
- Generate Offer > Pick template > Configure your package
- Under 'Net Salary' click Details

The basic salary should be 0 as no work hours have been recorded, note that this is different from having leaves recorded we're talking about a case where no records are available to compute the payslip basic salary. We might still want to generate a payslip in such cases to account for the flat allowances / deductions the employee might have on their contract's salary structure.

We can't just set it to 0 though:
This workaround is needed because the salary configurator also uses salary computation, where we do want to get the monthly wage (Since the contract is still provisional we don't have worked hours, but we still want it to be reflective of the position's monthly wage). In the case of an active employee however, paying a basic wage when no hours have been worked does not make sense so it should invariably be 0.

This could also be relevant with other work entry sources than attendances but a contract based on worked entries automatically generates worked hours according to the schedule so it is more difficult to reach.

opw-4266880

closes odoo/enterprise#73765

X-original-commit: 69338098f5407a01b4ddcb8b18bc82cf9d7cb4d5
Signed-off-by: Jurgen Gjini (jugj) <jugj@odoo.com>
Signed-off-by: Yoann Baron (bary) <bary@odoo.com>