[IMP] l10n_lu_hr_payroll: Update meal voucher computation on payslip.
An employee contributes €2.8 per day towards meal
vouchers. The total of €2.8 multiplied by the number
of workdays is either subtracted from the employee's
net salary or added to their gross salary as a
benefit in kind (BIK).

This amount is then considered part of the taxable
and contributory base.

Currently, only the 'Subtracted from net' scenario
is implemented.

Additionally, if the daily meal voucher value exceeds
the maximum limit (e.g., €15 for 2024), the
difference between the daily voucher amount and
this maximum should also be treated as BIK, similar
to the case where the employee's meal voucher
contribution is added to the gross salary.

This commit:
1. Fixes a bug in the 'Subtracted from net' case,
where the full meal voucher amount was incorrectly
deducted from the employee's net salary, instead
of only the employee's contribution.
2. Adds handling for the case where €2.8 times the
number of workdays is added to the gross salary.
3. Implements handling for situations where the
daily meal voucher amount exceeds the set maximum.
4. Makes minor adjustments to the meal vouchers
section in the contract view.

task-4310359

closes odoo/enterprise#73474

Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>