[IMP] knowledge: save article when pressing `Ctrl` + `s`
When the user presses `Ctrl` + `s`, the browser prompts the user to
select a directory to save the current page in. Today, most text editors
use the `Ctrl` + `s` shortcut to save the current file.

This commit will add a new event listener for `Ctrl` + `s` that will
prevent the default browser behavior and save the user changes in db.

task-4191390

closes odoo/enterprise#70166

Signed-off-by: Damien Abeloos (abd) <abd@odoo.com>