[IMP] sale_subscription: allow customers to pause/resume subscription from portal
Customers can now pause their subscriptions from the portal if the subscription
plan allows it. The pause duration is limited to one month and is only
applicable for current invoice periods.

A resume option is also provided, enabling customers to restart their
subscriptions before the pause ends. Upon resumption, the next invoice date
is set to the current date, and invoicing continues as usual.

task:3414468

closes odoo/enterprise#73490

Signed-off-by: Arnaud Joset (arj) <arj@odoo.com>