[FIX] sign: test_archived_requests_dont_send_reminders
This commit aims to fix an issue with test_archived_requests_dont_send_reminders.

Previously the test was trying to archive a record by performing a direct write command,
but since we have a designated action for archiving sign requests, it should be used.

Another issue is that when we archive a sign request, its state is set to 'canceled',
and for canceled sign requests we cannont update the validity,
however the test was trying to update the validity (as a way of jumping to the future).
But this is not the correct way of doing it, since we shouldn't change the validity if state is canceled.

To fix the issue, a mock_datetime_and_now is introduced.

closes odoo/enterprise#73156

Task: 4294860
Signed-off-by: Ahmad Almaghraby (alah) <alah@odoo.com>