[FIX] web_studio: fix web_studio_main_and_rename tour
At the end of the tour, when we want to export all modifications, it
opens a modal from v18.0. But, as it is not allowed to do action on
element that's below a modal, the tour fails.
In this commit, we click on export button to test the export feature
and then close the modal before continuing the tour.

closes odoo/enterprise#73589

X-original-commit: 578049256c5c96d4cc9e16984da7a0c68249abe1
Signed-off-by: Bruno Boi (boi) <boi@odoo.com>
Signed-off-by: Pierre Pulinckx (pipu) <pipu@odoo.com>