[IMP] mrp_workorder: add lot/serial barcode scanning in shop floor
Allow the shop floor user to quickly register component consumption by
enabling the scanning of lot/serial numbers in the register component
step of the shop floor.

To avoid having to fetch all available lots to the frontend, we use a
backend action to find a corresponding lot and to create/update the
stock move lines.

Part-of: odoo/enterprise#75532
Related: odoo/odoo#190467
Related: odoo/upgrade#6926
Signed-off-by: Quentin Wolfs (quwo) <quwo@odoo.com>