[FIX] documents_spreadsheet: adapt trend line tests
Since o-spreadsheet/commit/9909fa98d data points are normalized to compute
the trend line.

Those 2 tests were computing a 2nd order polynomial trend line based
on 2 data points only, which doesn't really makes sense.

With only 2 points, the normalized labels become [0, 1].
The matrix is of the form

x1	x2
x1^2	x2^2

it gives

0	1
0	1

with [0, 1] as the labels.

This matrix is not invertible since the two lines are the same and
the determinant is 0.

closes odoo/enterprise#73585

Related: odoo/odoo#186769
Signed-off-by: Rémi Rahir (rar) <rar@odoo.com>